MINOR: cache: Expose function getGenId()
This will be useful for debugging or testing when using the standalone
build program. It will allow to reuse a genid persistent cache file, just
like it is done when running as API server. For memo, the genid cache
provides a way to keep a stable device id between nscan databases.