FEATURE: build: Allow to read a genid cache file
This patch is for the standalone build program, which is mostly used for
debugging purposes.

The genid cache is used by the API server to provide a way to keep stable
device ids between nscan databases. The patch provides a way to read an
existing genid cache file when using the standalone build program. The
genid cache file is only read, it won't be modified.