MINOR: logger: Fix node console messing up with colors
Do not use console in dump(), it resets the color after the first line,
for instance when an object gets printed over multiple lines. The result
is the same when using a custom console with a "color" option set false,
node would reset with the default color after first line.

The patch just uses util.inspect() and process.stderr.write() in place
of console.error when the whole data to print is colored. Technically
this should not impact performance because console uses util.inspect()
internally.