BUG/MINOR: swspecs: Do not read configuration file, use config.options
Options from the configuration file are already available in
the config.options object, let's use it instead of reading and
parsing the file because this was already done.

Moreover, that code ignores the case when the path of the
configuration file was given to the program from the command line
with the -f or --config option.